DPR-564 WIP added ExternalMovementService, ExternalMovementRepository and reading data from JSON

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportingmi/controller/ExternalMovementsController.kt
@@ -28,8 +28,8 @@ class ExternalMovementsController(val externalMovementService: ExternalMovementS
     description = "Gets a count of external movements (mocked)",
     security = [ SecurityRequirement(name = "bearer-jwt") ],
   )
-  fun stubbedExternalMovements(@RequestParam("selectedPage") selectedPage: Int,
-                               @RequestParam("pageSize") pageSize: Int,
+  fun stubbedExternalMovements(@RequestParam("selectedPage") selectedPage: Long,
+                               @RequestParam("pageSize") pageSize: Long,
                                @RequestParam("sortColumn") sortColumn: String,
                                @RequestParam("sortedAsc") sortedAsc: Boolean,
                                ): List<ExternalMovement> {
